Restructured app module
